requirements.txt
aiogram
telethon
python-dotenv
sqlalchemy
asyncpg
grpcio
grpcio-tools
aiocryptopay